LoggerEditor now works without getting a list of loggers from a LoggerContext.  It only ever uses a ContentProvider.  However, getting all of the Loggers to come through is still a work in progress.
Also cleaned up the project some.  There are still a few class files hanging around from the initial commit when I just committed the entire folder, including the bin/ directories.

Finally, wrote a test app for the LogbackAndroidUI that allows us to have another app running our ContentProvider.  This is essential for effectively testing the LoggerEditor.  For now, the constants for the ContentProvider are stored in the CpConstants.java file so we can access them in the LogbackAndroidUI application.  This will change later on since each app using our libraries will have to define its own unique URI for the content provider.